Extract TideConditions component
